requirements.txt
fastapi
uvicorn[standard]
open_clip_torch
Pillow
torch
python-multipart
